refactor(auth): JWT 및 Security 관련 클래스 패키지 구조 정리
